requirements.txt
# Google Chat API dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# MCP Server
fastmcp>=0.1.0

# Configuration
pyyaml>=6.0

# Development/Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
